improve error reporting in PackageSet#normalize_vcs_list
The error messages now contain the section name and file as well as
the index of the offending entry. Moreover, this commit removes the
handling of idiosyncracies of Syck, Ruby's default YAML parser previous
to ruby 1.9 - which got removed on ruby 2.0. Psych (the current parser)
is now emitting errors on these weird constructs.